int to float

+++ Source/moorestech_client/Network/Send/AssemblePlayerPositionProtocol.cpp
@@ -5,7 +5,7 @@
 
 #include "moorestech_client/Network/Util/TArrayAddRange.h"
 
-TArray<uint8> UAssemblePlayerPositionProtocol::AssemblePlayerPosition(int x, int y,int playerId)
+TArray<uint8> UAssemblePlayerPositionProtocol::AssemblePlayerPosition(float x, float y,int playerId)
 {
 	TArray<uint8> assembleData;
 	TArrayAddRange::AddRange(assembleData,(int16)1);
